Code Check In Fot IsControlHeadEnum

+++ Coditech.Project/API/Comman/Coditech.API.Data/DataModel/Accounts/AccSetupGL.cs
@@ -22,7 +22,7 @@ namespace Coditech.API.Data
         public int PrintingSequence { get; set; }
         public bool IsActive { get; set; }
         public bool IsSystemGenerated { get; set; }
-        public  Nullable<byte> IsControlHeadEnum { get; set; }
+        public  Nullable<short> IsControlHeadEnum { get; set; }
         public Nullable<long> CreatedBy { get; set; }
         public Nullable<System.DateTime> CreatedDate { get; set; }
         public Nullable<long> ModifiedBy { get; set; }
